Move atom creation into base class
