+++ integration-tests/src/test/scala/IntegrationTestBase.scala
@@ -1,15 +1,18 @@
 package integration
 
+import java.time.Instant
+import java.util.UUID
+
 import com.google.api.client.googleapis.auth.oauth2.GoogleCredential
 import com.google.api.client.http.javanet.NetHttpTransport
 import com.google.api.client.json.jackson2.JacksonFactory
 import com.google.api.services.youtube.YouTube
+import com.gu.media.util.TestFilters
+import integration.services.{Config, GuHttp, TestAtomJsonGenerator}
 import org.scalatest.concurrent.{Eventually, IntegrationPatience}
 import org.scalatest.{BeforeAndAfterAll, FunSuite, Matchers}
-import integration.services.{Config, GuHttp, TestAtomJsonGenerator}
 import play.api.Logger
-
-import scala.collection.mutable.ListBuffer
+import play.api.libs.json.Json
 
 class IntegrationTestBase extends FunSuite with Matchers with Eventually with IntegrationPatience with GuHttp with TestAtomJsonGenerator with BeforeAndAfterAll {
 
@@ -17,42 +20,63 @@ class IntegrationTestBase extends FunSuite with Matchers with Eventually with In
 
   def apiUri(atomId: String): String = s"$targetBaseUrl/api/atom/$atomId"
 
-  val atomStore = new ListBuffer[String]() /* Add all created atoms IDs to this list as first action after atom created. This allows for test cleanup outside the test flow  */
-  val youtubeStore = new ListBuffer[String]() /* As above but for YouTube videos upload */
+  // For clean-up after tests
+  private var atomIds = List.empty[String]
+  private var youTubeIds = List.empty[String]
 
-  def deleteAtom(id: String) = {
-    Logger.info(s"Deleting atom $id")
-    gutoolsDelete(s"$targetBaseUrl/api2/atom/$id")
+  def createAtom(): String = {
+    val json = generateJson(
+      title = s"${TestFilters.testAtomBaseName}-${UUID.randomUUID().toString}",
+      description = "test atom",
+      category = "News",
+      channelId = Config.channelId,
+      youtubeCategoryId = Config.youtubeCategoryId,
+      expiryDate = Instant.now.toEpochMilli + (100 * 60 * 60 * 24)
+    )
+
+    val response = gutoolsPost(s"$targetBaseUrl/api2/atoms", jsonBody(json))
+
+    response.code() should be(201)
+
+    val atomId = (Json.parse(response.body().string()) \ "id").get.as[String]
+
+    Logger.info(s"Adding $atomId to Atom Store")
+    atomIds :+= atomId
+    atomId
   }
 
-  def deleteYouTubeVideo(id: String, client: YouTube) = {
-    Logger.info(s"Deleting YouTube video $id")
+  def addYouTubeVideoToStore(id: String): Unit = {
+    youTubeIds :+= id
+  }
 
-    client.videos()
-      .delete(id)
-      .setOnBehalfOfContentOwner(Config.youTube.contentOwner)
-      .execute()
+  override def beforeAll(): Unit = {
+    val aliveResponse = gutoolsGet(targetBaseUrl)
+    aliveResponse.code() should be (200)
+    aliveResponse.body().string() should include ("video")
+
+    super.beforeAll()
   }
 
   override def afterAll(): Unit = {
-    atomStore.foreach{ e => deleteAtom(e) }
+    atomIds.foreach { id =>
+      Logger.info(s"Deleting atom $id")
+      gutoolsDelete(s"$targetBaseUrl/api2/atom/$id")
+    }
 
-    if(youtubeStore.nonEmpty) {
+    if(youTubeIds.nonEmpty) {
       val client = youTubeClient()
-      youtubeStore.foreach(deleteYouTubeVideo(_, client))
+
+      youTubeIds.foreach { id =>
+        Logger.info(s"Deleting YouTube video $id")
+
+        client.videos()
+          .delete(id)
+          .setOnBehalfOfContentOwner(Config.youTube.contentOwner)
+          .execute()
+      }
     }
 
     super.afterAll()
-  }
-
-  def addAtomToStore(atomId: String): ListBuffer[String] = {
-    Logger.info(s"Adding $atomId to Atom Store")
-    atomStore += atomId
-  }
-
-  def addYouTubeVideoToStore(id: String) = {
-    Logger.info(s"Adding $id to YouTube Store")
-    youtubeStore += id
   }
 
   private def youTubeClient(): YouTube = {
